Move krusty test
